Added stuff
Added some functionality:
presentationBox- now contains 'layer' objects which hold a list of bufferedIMages, it also has methods to paint/ erase and go back. The way this works is everytime a graphic is added to the presentation, an image is created an added to the layer's imageList. When the presentation is painted, it goes through the layerlist and paints it's current position in it's imageList. Going back or forward just change a  layers' position in it's imagelist. Erasing takes the current image of a layer and removes pixels from where mouse is dragged. Note: it doesn't just paint those to white/ background color, this is so graphics can be removed from layers without removing from other layers. The annonmous layer class also holds a string name, this and the list of layers will be what constructs the graphicsBoxList. I have also added an eraser button and a go forward button.
Another thing I've added to the presentationbox is paintMode and eraserMode, this tells the mouseListeners what is added to the presentation image. These modes will also include an addShapesMode, addTextMode ect ect.