Color + brush size
very messy start to getting brush size and paint colour implemented. popupButton in the top right corner opens whatever mode's popup. paint popup has a slider to change brushsize and colorChooser to chnage paint color. Made an effort to get cursor to change to the size of the brush while it's being changed but it's messing up so will have to rethink how to do that.