Redesign + graphics box + media box
Done a huge redesign,changing the hierarchy completely, graphics box is the central control of pres elements on the slide. Graphics Box shows elements and can create a new graphics layer or delete a layer. A layer is selected by clicking it on, when a layer is selected tool bar is told and if the layer is a graphic specific layer then a mouse listener is added to the slide to add a graphics element. Graphic element contains an array of images (basically all the code that was previuously in the presentationBox). I've added a getter to the presElement to get a string of its type, i.e "TEXT" "GRAPHIC" "IMAGE" ect. so edditing presElements is easier. I've created the mediaBox, where when you initially click on it, it opens a filechooser than you can select a file. At the moment it only accepts ".jpeg" and ".png"  but this will be changed. - I had to create a new ImageElement class that doesn't have the url in the constructor. when an image is selected, an icon version of it appears in the media box, if you then click on the icon it then becomes the cursor and then if you click on the slide you want to add it to, the image origanlly selected is added to the slide as an Image element. The move button has also been slightly implemented, if a layer is selected and you click on the move icon, a 'move object' is diplayed around the presElement with arrow buttons in each corner to drag the to resize it. - this isnt fully finished but will finish after the whole scene has been finshed. I have also created three different layout types for the slide which are specified when the slide is created. (landscape, portrait, square). Have also added a background to the tabPane, idk i think it makes it look a bit more prefessional.