Refactoring
So the tabbedPane is the central class for instatiating slides, the components that need information about current visible slides are passed in through the constructor (elementsPanel + timeBar) or are instatiated within the class (tool bar+ mediaPanel).
Added a bunch of comments so should be fairly easy to understand whats happeneing.
I havn't finished this as theres an error whilst creating slides. Not entirely sure why they arent appearing in the tabPane, have tried everything i can think of to resolve it.
without the slide  appearing i cant make sure elements are being created right within elementsPanel or showing up correctly on elementsPropertiesPanel- will need some help to resolve this please
I have refactored so the ElementsPanel is the central class to store information about the presElements on the current visible slide, elementsProperties is passed in through the constructor